+++ debug/src/lib.rs
@@ -54,7 +54,7 @@ impl Plugin for DebugPlugin {
     fn build(&self, app: &mut AppBuilder) {
         app.add_resource(DebugMaterial::from(self.0))
             .init_resource::<DebugEntityMap>()
-            .add_stage_after(
+            .add_stage_before(
                 bevy_app::stage::POST_UPDATE,
                 "heron-debug",
                 SystemStage::serial(),