+++ CHANGELOG.md
@@ -19,6 +19,10 @@ The required rapier version is now >= 0.5.0 < 0.6.0.
 
 The variant `Body::Capsule` has been added. It takes a *haf_segment* and a *radius*.
 
+### Fixes
+
+* Misplaced debug render at startup
+
 
 ## [0.1.0-alpha.1] - 2021-01-30