+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is inspired from [Keep a Changelog], and this project adheres to [Sem
 
 ## [Unreleased]
 
+
+
+## [0.1.0] - 2021-02-15
+
 ### ⚠ Dependency requirement updated (breaking)
 
 The required rapier version is now >= 0.5.0 < 0.6.0.
@@ -86,5 +90,6 @@ One can read from `Events<CollisionEvent>` to be notified when collisions start 
 
 
 
-[Unreleased]: ../../compare/v0.1.0-alpha.1...HEAD
+[Unreleased]: ../../compare/v0.1.0...HEAD
+[0.1.0]: ../../compare/v0.1.0-alpha.1...v0.1.0
 [0.1.0-alpha.1]: ../../compare/...v0.1.0-alpha.1